backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
pymysql==1.1.1
python-dotenv==1.0.1
pillow==10.4.0
cryptography==42.0.8